refactor move monster action out of take monster turn action
