+++ src/features/monsters/actions/take-monsters-turn.js
@@ -1,6 +1,6 @@
 import attackPlayer    from './attack-player';
+import moveMonster     from './move-monster';
 import { SPRITE_SIZE } from '../../../config/constants';
-import getNextTile     from '../../../utils/get-next-tile';
 import { radiusTiles } from '../../../utils/get-surrounding-tiles';
 
 const MONSTER_ATTACK_RADIUS = 1;
@@ -52,23 +52,23 @@ export default function takeMonstersTurn() {
             // if the monster is mostly below the player on the y axis
             if(yDiff > 0) {
               // move the monster 'up' relatively
-              moveMonster('up', position, currentMap, id, 0, xDiff >= 0 ? 'left' : 'right');
+              dispatch(moveMonster('up', position, currentMap, id, 0, xDiff >= 0 ? 'left' : 'right'));
             }
             // if the monster is mostly above the player on the y axis
             else if(yDiff < 0) {
               // move the monster 'down' relatively
-              moveMonster('down', position, currentMap, id, 0, xDiff >= 0 ? 'left' : 'right');
+              dispatch(moveMonster('down', position, currentMap, id, 0, xDiff >= 0 ? 'left' : 'right'));
             }
           } else { // x axis is greater distance from player
             // if the monster is mostly to the right of the player
             if(xDiff > 0) {
               // move the monster 'left' relatively
-              moveMonster('left', position, currentMap, id, 0, yDiff >= 0 ? 'up' : 'down');
+              dispatch(moveMonster('left', position, currentMap, id, 0, yDiff >= 0 ? 'up' : 'down'));
             }
             // if the monster is mostly to the left of the player
             else if(xDiff < 0) {
               // move the monster 'right' relatively
-              moveMonster('right', position, currentMap, id, 0, yDiff >= 0 ? 'up' : 'down');
+              dispatch(moveMonster('right', position, currentMap, id, 0, yDiff >= 0 ? 'up' : 'down'));
             }
           }
         }
@@ -82,148 +82,10 @@ export default function takeMonstersTurn() {
         if((Math.round(Math.random() * (4 - 1) + 1)) !== 4) {
           let randomDirection = getRandomDirection();
           // move the monster in a random direction
-          moveMonster(randomDirection, position, currentMap, id, 0);
+          dispatch(moveMonster(randomDirection, position, currentMap, id, 0));
         }
       }
     });
-
-    function checkForOtherMonster(id, position, currentMap) {
-      // get current monsters
-      let monsterList = getState().monsters.components[currentMap];
-      let foundMonster = false;
-      // check list of monsters
-      Object.keys(monsterList).forEach(monsterId => {
-        // see if there's another monster in the next position
-        if(JSON.stringify(monsterList[monsterId].position) === JSON.stringify(position)) {
-          if(monsterId !== id) {
-            foundMonster = true;
-          }
-        }
-      })
-
-      return foundMonster;
-    }
-
-    function observeImpassable(newPos) {
-      const nextTile = getNextTile(getState().world, newPos);
-
-      return nextTile < 5 ? newPos : false;
-    }
-
-    // recursive function for moving the monster to the next available tile
-    // will try to go towards the player if possible
-    function moveMonster(direction, position, currentMap, id, count, preference = false) {
-      count ++;
-      // dont allow for infinite loops when monster can't move
-      if(count >= 5) return;
-
-      let nextPos = [0,0];
-
-      switch(direction) {
-        case 'up':
-          nextPos = [position[0], position[1] - SPRITE_SIZE];
-          // see if the monster can move to the next location
-          if(observeImpassable(nextPos)) {
-            // if we found a monster
-            if(checkForOtherMonster(id, nextPos, currentMap)) {
-              // move in a circle, but the opposite direction
-              return moveMonster(preference ? preference : 'left', position, currentMap, id, count);
-            } else {
-              // otherwise just move to the next spot
-              position[1] -= SPRITE_SIZE;
-            }
-            break;
-          } else {
-            // otherwise move them to another spot
-            return moveMonster(preference ? preference : 'right', position, currentMap, id, count);
-          }
-        case 'down':
-          nextPos = [position[0], position[1] + SPRITE_SIZE];
-          // see if the monster can move to the next location
-          if(observeImpassable(nextPos)) {
-            // if we found a monster
-            if(checkForOtherMonster(id, nextPos, currentMap)) {
-              // move in a circle, but the opposite direction
-              return moveMonster(preference ? preference : 'right', position, currentMap, id, count);
-            } else {
-              // otherwise just move to the next spot
-              position[1] += SPRITE_SIZE;
-            }
-            break;
-          } else {
-            // otherwise move them to another spot
-            return moveMonster(preference ? preference : 'left', position, currentMap, id, count);
-          }
-        case 'left':
-          nextPos = [position[0] - SPRITE_SIZE, position[1]];
-          // see if the monster can move to the next location
-          if(observeImpassable(nextPos)) {
-            // if we found a monster
-            if(checkForOtherMonster(id, nextPos, currentMap)) {
-              // move in a circle, but the opposite direction
-              return moveMonster(preference ? preference : 'down', position, currentMap, id, count);
-            } else {
-              // otherwise just move to the next spot
-              position[0] -= SPRITE_SIZE;
-            }
-            break;
-          } else {
-            // otherwise move them to another spot
-            return moveMonster(preference ? preference : 'up', position, currentMap, id, count);
-          }
-        case 'right':
-          nextPos = [position[0] + SPRITE_SIZE, position[1]];
-          // see if the monster can move to the next location
-          if(observeImpassable(nextPos)) {
-            // if we found a monster
-            if(checkForOtherMonster(id, nextPos, currentMap)) {
-              // move in a circle, but the opposite direction
-              return moveMonster(preference ? preference : 'up', position, currentMap, id, count);
-            } else {
-              // otherwise just move to the next spot
-              position[0] += SPRITE_SIZE;
-            }
-            break;
-          } else {
-            // otherwise move them to another spot
-            return moveMonster(preference ? preference : 'down', position, currentMap, id, count);
-          }
-        default:
-      }
-      // recalculate if the monster is in sight
-      const { sightBox } = getState().map;
-      let inSight = false;
-      // look through each current sight box tile
-      sightBox.forEach(tile => {
-        // if the monster is in sight
-        const newMonsterPos = position.map(value => value / SPRITE_SIZE);
-        if(JSON.stringify(tile) === JSON.stringify(newMonsterPos)) {
-          inSight = true;
-        }
-      });
-      // if the monster is now in sight
-      if(inSight) {
-        dispatch({
-          type: 'REVEAL_MONSTER',
-          payload: { id, map: currentMap }
-        });
-      } else {
-        // if the monster is now out of sight
-        dispatch({
-          type: 'HIDE_MONSTER',
-          payload: { id, map: currentMap }
-        });
-      }
-      // move the monster
-      dispatch({
-        type: 'MOVE_MONSTER',
-        payload: {
-          map: currentMap,
-          id,
-          position
-        }
-      });
-    }
   }
 }
